# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "IMF Gadget API",
    "description": "API for managing gadgets in the IMF."
  },
  "paths": {
    "/gadgets": {
      "get": {
        "summary": "Get all gadgets (with optional status filter)",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["Available", "Deployed", "Destroyed", "Decommissioned"]
            },
            "description": "Filter gadgets by status"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of gadgets",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Gadget"
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Add a new gadget",
        "parameters": [
          {
            "name": "gadget",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Gadget"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Gadget added"
          }
        }
      }
    },
    "/gadgets/{id}": {
      "get": {
        "summary": "Get a specific gadget",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "A gadget object",
            "schema": {
              "$ref": "#/definitions/Gadget"
            }
          },
          "404": {
            "description": "Gadget not found"
          }
        }
      },
      "patch": {
        "summary": "Update a specific gadget",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GadgetUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated gadget object",
            "schema": {
              "$ref": "#/definitions/Gadget"
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "500": {
            "description": "Failed to update gadget"
          }
        }
      },
      "delete": {
        "summary": "Delete a specific gadget",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Gadget deleted successfully"
          },
          "404": {
            "description": "Gadget not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Gadget": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      },
      "required": ["id", "name", "status"]
    },
    "GadgetUpdate": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "enum": ["Available", "Deployed", "Destroyed", "Decommissioned"]
        }
      },
      "required": ["name", "status"]
    }
  }
}